# Speed up block serialization