--- a/x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/data/ElementType.java
+++ b/x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/data/ElementType.java
@@ -8,37 +8,46 @@
 package org.elasticsearch.compute.data;
 
 import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.TransportVersions;
+import org.elasticsearch.common.io.stream.StreamInput;
+import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.common.util.Maps;
+
+import java.io.IOException;
+import java.util.Map;
 
 /**
  * The type of elements in {@link Block} and {@link Vector}
  */
 public enum ElementType {
-    BOOLEAN("Boolean", BlockFactory::newBooleanBlockBuilder),
-    INT("Int", BlockFactory::newIntBlockBuilder),
-    LONG("Long", BlockFactory::newLongBlockBuilder),
-    FLOAT("Float", BlockFactory::newFloatBlockBuilder),
-    DOUBLE("Double", BlockFactory::newDoubleBlockBuilder),
+    BOOLEAN(0, "Boolean", BlockFactory::newBooleanBlockBuilder, BooleanBlock::readFrom),
+    INT(1, "Int", BlockFactory::newIntBlockBuilder, IntBlock::readFrom),
+    LONG(2, "Long", BlockFactory::newLongBlockBuilder, LongBlock::readFrom),
+    FLOAT(3, "Float", BlockFactory::newFloatBlockBuilder, FloatBlock::readFrom),
+    DOUBLE(4, "Double", BlockFactory::newDoubleBlockBuilder, DoubleBlock::readFrom),
     /**
      * Blocks containing only null values.
      */
-    NULL("Null", (blockFactory, estimatedSize) -> new ConstantNullBlock.Builder(blockFactory)),
+    NULL(5, "Null", (blockFactory, estimatedSize) -> new ConstantNullBlock.Builder(blockFactory), BlockStreamInput::readConstantNullBlock),
 
-    BYTES_REF("BytesRef", BlockFactory::newBytesRefBlockBuilder),
+    BYTES_REF(6, "BytesRef", BlockFactory::newBytesRefBlockBuilder, BytesRefBlock::readFrom),
 
     /**
      * Blocks that reference individual lucene documents.
      */
-    DOC("Doc", DocBlock::newBlockBuilder),
+    DOC(7, "Doc", DocBlock::newBlockBuilder, in -> { throw new UnsupportedOperationException("can't read doc blocks"); }),
 
     /**
      * Composite blocks which contain array of sub-blocks.
      */
-    COMPOSITE("Composite", BlockFactory::newAggregateMetricDoubleBlockBuilder),
+    COMPOSITE(8, "Composite", BlockFactory::newAggregateMetricDoubleBlockBuilder, CompositeBlock::readFrom),
 
     /**
      * Intermediate blocks which don't support retrieving elements.
      */
-    UNKNOWN("Unknown", (blockFactory, estimatedSize) -> { throw new UnsupportedOperationException("can't build null blocks"); });
+    UNKNOWN(9, "Unknown", (blockFactory, estimatedSize) -> { throw new UnsupportedOperationException("can't build null blocks"); }, in -> {
+        throw new UnsupportedOperationException("can't read unknown blocks");
+    });
 
     private interface BuilderSupplier {
         Block.Builder newBlockBuilder(BlockFactory blockFactory, int estimatedSize);
@@ -47,9 +56,19 @@ public enum ElementType {
     private final String pascalCaseName;
     private final BuilderSupplier builder;
 
-    ElementType(String pascalCaseName, BuilderSupplier builder) {
+    private final byte writableCode;
+    private final String legacyWritableName;
+    final BlockReader reader;
+
+    ElementType(int writeableCode, String pascalCaseName, BuilderSupplier builder, BlockReader reader) {
+        this.writableCode = (byte) writeableCode;
+        if (this.writableCode != writeableCode) {
+            throw new IllegalArgumentException("code must be in the range [0, " + Byte.MAX_VALUE + "); got " + writeableCode);
+        }
         this.pascalCaseName = pascalCaseName;
+        this.legacyWritableName = "Null".equals(pascalCaseName) ? "ConstantNullBlock" : pascalCaseName + "Block";
         this.builder = builder;
+        this.reader = reader;
     }
 
     /**
@@ -85,5 +104,45 @@ public enum ElementType {
 
     public String pascalCaseName() {
         return pascalCaseName;
+    }
+
+    private static final Map<String, ElementType> fromLegacyNames;
+
+    static {
+        ElementType[] values = values();
+        fromLegacyNames = Maps.newMapWithExpectedSize(values.length);
+        for (int i = 0; i < values.length; i++) {
+            assert values[i].writableCode == i;
+            fromLegacyNames.put(values[i].legacyWritableName, values[i]);
+        }
+    }
+
+    /**
+     * Read element type from an input stream
+     */
+    static ElementType readFrom(StreamInput in) throws IOException {
+        if (in.getTransportVersion().onOrAfter(TransportVersions.ESQL_SERIALIZE_BLOCK_TYPE_CODE)) {
+            byte b = in.readByte();
+            return values()[b];
+        } else {
+            final String writeableName = in.readString();
+            ElementType elementType = fromLegacyNames.get(writeableName);
+            if (elementType == null) {
+                throw new IllegalArgumentException("Unknown element type for named writable [" + writeableName + "]");
+            }
+            return elementType;
+        }
+    }
+
+    void writeTo(StreamOutput out) throws IOException {
+        if (out.getTransportVersion().onOrAfter(TransportVersions.ESQL_SERIALIZE_BLOCK_TYPE_CODE)) {
+            out.writeByte(writableCode);
+        } else {
+            out.writeString(legacyWritableName);
+        }
+    }
+
+    interface BlockReader {
+        Block readBlock(BlockStreamInput in) throws IOException;
     }
 }