--- a/x-pack/plugin/esql/compute/src/test/java/org/elasticsearch/compute/data/SerializationTestCase.java
+++ b/x-pack/plugin/esql/compute/src/test/java/org/elasticsearch/compute/data/SerializationTestCase.java
@@ -7,17 +7,17 @@
 
 package org.elasticsearch.compute.data;
 
+import org.elasticsearch.TransportVersion;
 import org.elasticsearch.common.breaker.CircuitBreaker;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.io.stream.ByteBufferStreamInput;
 import org.elasticsearch.common.io.stream.BytesStreamOutput;
-import org.elasticsearch.common.io.stream.NamedWriteableAwareStreamInput;
-import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.common.util.BigArrays;
 import org.elasticsearch.common.util.MockBigArrays;
 import org.elasticsearch.common.util.PageCacheRecycler;
 import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.test.TransportVersionUtils;
 import org.junit.After;
 import org.junit.Before;
 
@@ -29,7 +29,6 @@ import static org.hamcrest.Matchers.equalTo;
 public abstract class SerializationTestCase extends ESTestCase {
     BigArrays bigArrays;
     protected BlockFactory blockFactory;
-    NamedWriteableRegistry registry = new NamedWriteableRegistry(BlockWritables.getNamedWriteables());
 
     @Before
     public final void newBlockFactory() {
@@ -41,7 +40,6 @@ public abstract class SerializationTestCase extends ESTestCase {
     public final void blockFactoryEmpty() {
         assertThat(blockFactory.breaker().getUsed(), equalTo(0L));
         blockFactory = null;
-        registry = null;
     }
 
     Page serializeDeserializePage(Page origPage) throws IOException {
@@ -52,18 +50,18 @@ public abstract class SerializationTestCase extends ESTestCase {
     }
 
     BlockStreamInput blockStreamInput(BytesStreamOutput out) {
-        return new BlockStreamInput(
-            new NamedWriteableAwareStreamInput(ByteBufferStreamInput.wrap(BytesReference.toBytes(out.bytes())), registry),
-            blockFactory
-        );
+        return new BlockStreamInput(ByteBufferStreamInput.wrap(BytesReference.toBytes(out.bytes())), blockFactory);
     }
 
     @SuppressWarnings("unchecked")
     <T extends Block> T serializeDeserializeBlock(T origBlock) throws IOException {
         try (BytesStreamOutput out = new BytesStreamOutput()) {
-            out.writeNamedWriteable(origBlock);
+            TransportVersion version = TransportVersionUtils.randomCompatibleVersion(random());
+            out.setTransportVersion(version);
+            Block.writeTypedBlock(origBlock, out);
             try (BlockStreamInput in = blockStreamInput(out)) {
-                return (T) in.readNamedWriteable(Block.class);
+                in.setTransportVersion(version);
+                return (T) Block.readTypedBlock(in);
             }
         }
     }