Currently, we use `NamedWriteable` for serializing blocks. While convenient, it incurs a noticeable performance penalty when pages contain thousands of blocks. Since block types are small and already centered in `ElementType`, we can safely switch from `NamedWriteable` to typed code. For example, the `NamedWriteable` alone of a small page with 10K fields would be 180KB, whereas the new method reduces it to 10KB. Below are the serialization improvements with `FROM idx | LIMIT 10000` where the target index has 10K fields:

- write_exchange_response executed 173 times took: 73.2ms -> 26.7ms
- read_exchange_response executed 173 times took: 49.4ms ->  25.8ms

---
- I might open another PR to avoid serializing `positionCount` as we should already have it from the page. 
- We might need to do the same thing for plan serialization.